reviews, ratings and API security done

+++ controllers/courses.js
@@ -14,7 +14,7 @@ exports.getCourses = asyncHandler(async (req, res, next) => {
     return res.status(200).json({
       success: true,
       count: courses.length,
-      data: courses
+      data: courses,
     });
   } else {
     res.status(200).json(res.advancedResults);
@@ -47,22 +47,33 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @route     POST /api/v1/bootcamps/:bootcampId/courses
 // @access    Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
-    req.body.bootcamp = req.params.bootcampId;
-  
-    const bootcamp = await Bootcamp.findById(req.params.bootcampId);
-  
-    if (!bootcamp) {
-      return next(
-        new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`),
-        404
-      );
-    }
+  req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
+
+  const bootcamp = await Bootcamp.findById(req.params.bootcampId);
+
+  if (!bootcamp) {
+    return next(
+      new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`),
+      404
+    );
+  }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to ${bootcamp._id}`,
+        401
+      )
+    );
+  }
 
   const course = await Course.create(req.body);
 
   res.status(200).json({
     success: true,
-    data: course
+    data: course,
   });
 });
 
@@ -70,23 +81,33 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 // @route     PUT /api/v1/courses/:id
 // @access    Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-    let course = await Course.findById(req.params.id);
-  
-    if (!course) {
-      return next(
-        new ErrorResponse(`No course with the id of ${req.params.id}`),
-        404
-      );
-    }
+  let course = await Course.findById(req.params.id);
 
-    course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    })
+  if (!course) {
+    return next(
+      new ErrorResponse(`No course with the id of ${req.params.id}`),
+      404
+    );
+  }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update a course to ${course._id}`,
+        401
+      )
+    );
+  }
+
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
 
   res.status(200).json({
     success: true,
-    data: course
+    data: course,
   });
 });
 
@@ -94,19 +115,29 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @route     DELETE /api/v1/courses/:id
 // @access    Private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
-    const course = await Course.findById(req.params.id);
-  
-    if (!course) {
-      return next(
-        new ErrorResponse(`No course with the id of ${req.params.id}`),
-        404
-      );
-    }
+  const course = await Course.findById(req.params.id);
 
-    await course.deleteOne();
+  if (!course) {
+    return next(
+      new ErrorResponse(`No course with the id of ${req.params.id}`),
+      404
+    );
+  }
+
+   // Make sure user is course owner
+   if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete a course to ${course._id}`,
+        401
+      )
+    );
+  }
+
+  await course.deleteOne();
 
   res.status(200).json({
     success: true,
-    data: {}
+    data: {},
   });
 });
